Update parser_project.py
I built this for my colleagues. We had a very large spreadsheet to navigate and this cut downtime effectively.  Essentially, if someone had an inquiry, for example, a problem with the app, I could type 'app' and the data frame will return all rows and columns associated with 'app' (not case senstative). If you plan to use this, you may need to tag a column with whatever you plan to search for.  For example, we had app - not working, app - enhancement, etc... each of these has a different value in the column. Create a new column and label each row in the new column app by app - not working for example and the spreadsheet will pull everything associated with that tag.